Add files via upload
The Functions document contains all the real parameters used to simulate the system. Those are the orbital velocities, masses and distances to sun of all the planets from the solar system. We apply a reescalation of time, distance and mass. All the distances are divided by the sun-earth distance. Masses are reescalated by the sun mass. 

In the main document we apply the Verlet algorithm and create an animation of the planets motion. We check the simulation is correct by measuring the periods of the planets in the simulation and checking they agree with the real periods.